refactor: streamline secret shares and policy utility functions
