+++ crates/tessera-abe/src/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod aes;
 pub mod secret_shares;
-pub mod tools;
